cryptarchia: move mk_genesis_state to test_common
